% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMember.R
\name{addMember}
\alias{addMember}
\title{Agregar un user a Ghost}
\usage{
addMember(name, email)
}
\arguments{
\item{name}{El nombre del usuario}

\item{email}{El email del usuario}
}
\value{
Devuelve un tibble con los siguientes campos:
- status El código de status de la respuesta
- memberID El ID del miembro creado
}
\description{
Esta función agrega un usuario a Ghost.
}
\examples{
addMember("Juan Perez", "juan@perez.com"

}
