% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getJWT.R
\name{getJWT}
\alias{getJWT}
\title{Crea un JWT para integración con Ghost.or}
\usage{
getJWT(integration = "admin")
}
\arguments{
\item{integration}{Qué variable tomará. Default es "admin"}
}
\value{
Devuelve un JWT para integración con Ghost
}
\description{
Toma el secret de la variable de entorno GHOST_SECRET_ADMIN
y crea un JWT para integración con Ghost.
}
\details{
El secret se obtiene de la variable de entorno GHOST_SECRET_ADMIN
que se setea en el archivo .Renviron
el nombre en el archivo es GHOST_SECRET_ADMIN o
GHOST_SECRET_[USER]
La función toma el valor de la variable de entorno y lo divide en dos partes
separadas por ":". La primera parte es el kid y la segunda el secret.
}
\examples{
getJWT("admin") -> Devuelve un JWT para integración con Ghost
}
