% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMembers.R
\name{getMembers}
\alias{getMembers}
\title{Devuelve miembros de Ghost}
\usage{
getMembers(requestPage = 0, integration = "admin")
}
\arguments{
\item{requestPage}{El número de página que se quiere consultar. Cada página
devuelve 15 registros.
Si se utiliza 0, siendo también el valor por default, devuelve todos los registros.
Utilizar con precaución.}
}
\value{
Devuelve un tibble con los siguientes campos:
- name
- email
- id         El ID interno usado por Ghost. Este sirve para consulta específica
- uuid       Adicional a id
- suscribed  Si tiene suscripciones o no a newsletters. No lo uso aún
- status     Acá todos dirán free. Pero lo dejo para un futuro.
}
\description{
'reqGhostMembersPaged' devuelve los miembros en páginas de a 15
dado que así son las especificaciones de Ghost.
}
\details{
La API de Ghost pagina en multiplos de 15 (la última puede contener)
menos elementos.
Esta función toma el parámetro requestPage, seteado como default en 0, 
y ese parámetro lo pasa a la API para pedirle la página en cuestión
}
\examples{
reqGhostMembersPaged(15) -> Devuelve la página 15
}
